requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==40.0.2
Django==4.2.8
djangorestframework==3.14.0
ecdsa==0.18.0
mysqlclient==2.2.1
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
